wip tests but one pass, generalized some code
