+++ packages/algjulia-service/Project.toml
@@ -18,6 +18,7 @@ JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MLStyle = "d8e11817-5142-5d16-987a-aa16d5891078"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
@@ -32,4 +33,5 @@ JSON3 = "1"
 LinearAlgebra = "1"
 MLStyle = "0.4"
 OrdinaryDiffEq = "6"
+Reexport = "1.2.2"
 StaticArrays = "1"
