+++ packages/algjulia-service/test/test_jsons/diffusion_data_twovars.json
@@ -4,7 +4,8 @@
   "initialConditions": {
 	"01932402-bcf5-7432-8d14-dbae9eabf907": "Gaussian"
   },
-  "plotVariables": ["01932402-bcf5-7432-8d14-dbae9eabf907"],
+  "scalars": {},
+  "plotVariables": ["01932402-bcf5-7432-8d14-dbae9eabf907","01932403-5b6c-7231-90d7-d7cece275eb2"],
   "diagram": [
     {
       "id": "01932402-bcf5-7432-8d14-dbae9eabf907",
